file field; form element updates

+++ src/Fieldtypes/FileFieldtype.php
@@ -76,49 +76,51 @@ class FileFieldtype extends Fieldtype
      */
     public function persistRelationship($model, Field $field)
     {
-        $files     = request()->file($field->handle);
-        $directory = Directory::firstOrCreate([
-            'name' => ($name = $field->settings['directory'] ?? 'uploads'),
-            'slug' => Str::slug($name),
-        ]);
+        if (request()->hasFile($field->handle)) {
+            $files     = request()->file($field->handle);
+            $directory = Directory::firstOrCreate([
+                'name' => ($name = $field->settings['directory'] ?? 'uploads'),
+                'slug' => Str::slug($name),
+            ]);
 
-        $oldValues = $model->{$field->handle}->pluck('id');
-        $newValues = collect($files)
-            ->mapWithKeys(function($file, $key) use ($field, $directory) {
-                $uuid      = unique_id();
-                $name      = pathinfo($file->getClientOriginalName(), PATHINFO_FILENAME);
-                $extension = $file->extension();
-                $bytes     = $file->getSize();
-                $mimetype  = $file->getClientMimeType();
-                $filetype  = strtok($mimetype, '/');
-                $location  = "files/{$uuid}-{$name}.{$extension}";
+            $oldValues = $model->{$field->handle}->pluck('id');
+            $newValues = collect($files)
+                ->mapWithKeys(function($file, $key) use ($field, $directory) {
+                    $uuid      = unique_id();
+                    $name      = pathinfo($file->getClientOriginalName(), PATHINFO_FILENAME);
+                    $extension = $file->extension();
+                    $bytes     = $file->getSize();
+                    $mimetype  = $file->getClientMimeType();
+                    $filetype  = strtok($mimetype, '/');
+                    $location  = "files/{$uuid}-{$name}.{$extension}";
 
-                Storage::disk('public')->putFileAs('', $file, $location);
+                    Storage::disk('public')->putFileAs('', $file, $location);
 
-                if ($filetype == 'image') {
-                    list($width, $height) = getimagesize($file);
-                }
+                    if ($filetype == 'image') {
+                        list($width, $height) = getimagesize($file);
+                    }
 
-                $model = FileModel::create([
-                    'directory_id' => $directory->id,
-                    'uuid'         => $uuid,
-                    'name'         => $name,
-                    'extension'    => $extension,
-                    'bytes'        => $bytes,
-                    'mimetype'     => $mimetype,
-                    'location'     => $location,
-                    'width'        => $width ?? null,
-                    'height'       => $height ?? null,
-                ]);
+                    $model = FileModel::create([
+                        'directory_id' => $directory->id,
+                        'uuid'         => $uuid,
+                        'name'         => $name,
+                        'extension'    => $extension,
+                        'bytes'        => $bytes,
+                        'mimetype'     => $mimetype,
+                        'location'     => $location,
+                        'width'        => $width ?? null,
+                        'height'       => $height ?? null,
+                    ]);
 
-                return [ $model['id'] => [
-                    'field_id' => $field->id,
-                    'order'    => $key + 1
-                ]];
-            });
+                    return [ $model['id'] => [
+                        'field_id' => $field->id,
+                        'order'    => $key + 1
+                    ]];
+                });
 
-        $model->{$field->handle}()->detach($oldValues);
-        $model->{$field->handle}()->attach($newValues);
+            $model->{$field->handle}()->detach($oldValues);
+            $model->{$field->handle}()->attach($newValues);
+        }
     }
 
     /**
